requirements.txt
# Core ML and RAG Dependencies
llama_index>=0.9.0
langchain>=0.1.0
transformers>=4.38.0
torch>=2.2.0
sentence-transformers>=2.5.0

# LLM Specific Dependencies (CPU Friendly)
ctransformers>=0.2.27
llama-cpp-python>=0.2.0
accelerate>=0.26.0

# Web Application
streamlit>=1.32.0
streamlit-extras>=0.3.6
python-dotenv>=1.0.0

# Data Processing & PDF Handlingf
pandas>=2.2.0
numpy>=1.26.0
pypdf>=4.0.0
pdfplumber>=0.10.3

# Vector Store
chromadb>=0.4.22

# Additional Dependencies
langchain-community>=0.0.10
tqdm>=4.66.0

# Development
# -e .